Added new general utils.js to store functions in. Added chargetime function to get calculated chargetimeleft and finishtime from everywhere and by one single function => deduplication of functions.
